+++ examples/resources/leaseweb_public_cloud_ip/import.sh
@@ -1,0 +1,2 @@
+# Public Cloud ip can be imported by specifying <instance_id>,<ip>.
+terraform import leaseweb_public_cloud_ip.example 695ddd91-051f-4dd6-9120-938a927a47d0,10.0.0.1