+++ internal/provider/publiccloud/instance_resource_test.go
@@ -55,7 +55,7 @@ func generateContractObject(
 
 	contract, _ := types.ObjectValueFrom(
 		context.TODO(),
-		contractResourceModel{}.AttributeTypes(),
+		contractResourceModel{}.attributeTypes(),
 		contractResourceModel{
 			BillingFrequency: basetypes.NewInt32Value(int32(*billingFrequency)),
 			Term:             basetypes.NewInt32Value(int32(*contractTerm)),
@@ -72,7 +72,7 @@ func generateInstanceResourceModel() instanceResourceModel {
 
 	image, _ := types.ObjectValueFrom(
 		context.TODO(),
-		imageResourceModel{}.AttributeTypes(),
+		imageResourceModel{}.attributeTypes(),
 		imageResourceModel{
 			ID:           basetypes.NewStringValue("UBUNTU_20_04_64BIT"),
 			MarketApps:   emptyList,
@@ -157,14 +157,14 @@ func Test_adaptInstanceDetailsToInstanceResource(t *testing.T) {
 	assert.Equal(t, "isoId", iso.ID.ValueString())
 }
 
-func Test_instanceResourceModel_GetLaunchInstanceOpts(t *testing.T) {
+func Test_instanceResourceModel_getLaunchOpts(t *testing.T) {
 	t.Run("required values are set", func(t *testing.T) {
 		instance := generateInstanceResourceModel()
 		instance.MarketAppID = basetypes.NewStringPointerValue(nil)
 		instance.Reference = basetypes.NewStringPointerValue(nil)
 		instance.RootDiskSize = basetypes.NewInt32PointerValue(nil)
 
-		got, err := instance.GetLaunchInstanceOpts(context.TODO())
+		got, err := instance.getLaunchOpts(context.TODO())
 
 		assert.NoError(t, err)
 		assert.Equal(t, publiccloud.REGIONNAME_EU_WEST_3, got.Region)
@@ -188,7 +188,7 @@ func Test_instanceResourceModel_GetLaunchInstanceOpts(t *testing.T) {
 	t.Run("optional values are passed", func(t *testing.T) {
 		instance := generateInstanceResourceModel()
 
-		got, err := instance.GetLaunchInstanceOpts(context.TODO())
+		got, err := instance.getLaunchOpts(context.TODO())
 
 		assert.NoError(t, err)
 		assert.Equal(t, "marketAppId", *got.MarketAppId)
@@ -202,7 +202,7 @@ func Test_instanceResourceModel_GetLaunchInstanceOpts(t *testing.T) {
 			instance := generateInstanceResourceModel()
 			instance.RootDiskStorageType = basetypes.NewStringValue("tralala")
 
-			_, err := instance.GetLaunchInstanceOpts(context.TODO())
+			_, err := instance.getLaunchOpts(context.TODO())
 
 			assert.Error(t, err)
 			assert.ErrorContains(t, err, "tralala")
@@ -215,7 +215,7 @@ func Test_instanceResourceModel_GetLaunchInstanceOpts(t *testing.T) {
 			instance := generateInstanceResourceModel()
 			instance.Type = basetypes.NewStringValue("tralala")
 
-			_, err := instance.GetLaunchInstanceOpts(context.TODO())
+			_, err := instance.getLaunchOpts(context.TODO())
 
 			assert.Error(t, err)
 			assert.ErrorContains(t, err, "tralala")
@@ -230,7 +230,7 @@ func Test_instanceResourceModel_GetLaunchInstanceOpts(t *testing.T) {
 			contract := generateContractObject(nil, nil, &contractType)
 			instance.Contract = contract
 
-			_, err := instance.GetLaunchInstanceOpts(context.TODO())
+			_, err := instance.getLaunchOpts(context.TODO())
 
 			assert.Error(t, err)
 			assert.ErrorContains(t, err, "tralala")
@@ -245,7 +245,7 @@ func Test_instanceResourceModel_GetLaunchInstanceOpts(t *testing.T) {
 			contract := generateContractObject(nil, &contractTerm, nil)
 			instance.Contract = contract
 
-			_, err := instance.GetLaunchInstanceOpts(context.TODO())
+			_, err := instance.getLaunchOpts(context.TODO())
 
 			assert.Error(t, err)
 			assert.ErrorContains(t, err, "555")
@@ -260,7 +260,7 @@ func Test_instanceResourceModel_GetLaunchInstanceOpts(t *testing.T) {
 			contract := generateContractObject(&billingFrequency, nil, nil)
 			instance.Contract = contract
 
-			_, err := instance.GetLaunchInstanceOpts(context.TODO())
+			_, err := instance.getLaunchOpts(context.TODO())
 
 			assert.Error(t, err)
 			assert.ErrorContains(t, err, "555")
@@ -271,7 +271,7 @@ func Test_instanceResourceModel_GetLaunchInstanceOpts(t *testing.T) {
 		instance := generateInstanceResourceModel()
 		instance.Region = basetypes.NewStringValue("tralala")
 
-		_, err := instance.GetLaunchInstanceOpts(context.TODO())
+		_, err := instance.getLaunchOpts(context.TODO())
 
 		assert.Error(t, err)
 		assert.ErrorContains(t, err, "tralala")
@@ -283,7 +283,7 @@ func Test_instanceResourceModel_GetLaunchInstanceOpts(t *testing.T) {
 			instance := generateInstanceResourceModel()
 			instance.Image = basetypes.NewObjectNull(map[string]attr.Type{})
 
-			_, err := instance.GetLaunchInstanceOpts(context.TODO())
+			_, err := instance.getLaunchOpts(context.TODO())
 
 			assert.Error(t, err)
 			assert.ErrorContains(t, err, ".imageResourceModel")
@@ -296,7 +296,7 @@ func Test_instanceResourceModel_GetLaunchInstanceOpts(t *testing.T) {
 			instance := generateInstanceResourceModel()
 			instance.Contract = basetypes.NewObjectNull(map[string]attr.Type{})
 
-			_, err := instance.GetLaunchInstanceOpts(context.TODO())
+			_, err := instance.getLaunchOpts(context.TODO())
 
 			assert.Error(t, err)
 			assert.ErrorContains(t, err, ".contractResourceModel")
@@ -304,11 +304,11 @@ func Test_instanceResourceModel_GetLaunchInstanceOpts(t *testing.T) {
 	)
 }
 
-func Test_instanceResourceModel_GetUpdateInstanceOpts(t *testing.T) {
+func Test_instanceResourceModel_getUpdateOpts(t *testing.T) {
 	t.Run("optional values are set", func(t *testing.T) {
 		instance := generateInstanceResourceModel()
 
-		got, err := instance.GetUpdateInstanceOpts(context.TODO())
+		got, err := instance.getUpdateOpts(context.TODO())
 
 		assert.NoError(t, err)
 		assert.Equal(t, publiccloud.TYPENAME_M5A_4XLARGE, *got.Type)
@@ -325,7 +325,7 @@ func Test_instanceResourceModel_GetUpdateInstanceOpts(t *testing.T) {
 			instance := generateInstanceResourceModel()
 			instance.Type = basetypes.NewStringValue("tralala")
 
-			_, err := instance.GetUpdateInstanceOpts(context.TODO())
+			_, err := instance.getUpdateOpts(context.TODO())
 
 			assert.Error(t, err)
 			assert.ErrorContains(t, err, "tralala")
@@ -340,7 +340,7 @@ func Test_instanceResourceModel_GetUpdateInstanceOpts(t *testing.T) {
 			contract := generateContractObject(nil, nil, &contractType)
 			instance.Contract = contract
 
-			_, err := instance.GetUpdateInstanceOpts(context.TODO())
+			_, err := instance.getUpdateOpts(context.TODO())
 
 			assert.Error(t, err)
 			assert.ErrorContains(t, err, "tralala")
@@ -355,7 +355,7 @@ func Test_instanceResourceModel_GetUpdateInstanceOpts(t *testing.T) {
 			contract := generateContractObject(nil, &contractTerm, nil)
 			instance.Contract = contract
 
-			_, err := instance.GetUpdateInstanceOpts(context.TODO())
+			_, err := instance.getUpdateOpts(context.TODO())
 
 			assert.Error(t, err)
 			assert.ErrorContains(t, err, "555")
@@ -370,7 +370,7 @@ func Test_instanceResourceModel_GetUpdateInstanceOpts(t *testing.T) {
 			contract := generateContractObject(&billingFrequency, nil, nil)
 			instance.Contract = contract
 
-			_, err := instance.GetUpdateInstanceOpts(context.TODO())
+			_, err := instance.getUpdateOpts(context.TODO())
 
 			assert.Error(t, err)
 			assert.ErrorContains(t, err, "555")
@@ -383,51 +383,10 @@ func Test_instanceResourceModel_GetUpdateInstanceOpts(t *testing.T) {
 			instance := generateInstanceResourceModel()
 			instance.Contract = basetypes.NewObjectNull(map[string]attr.Type{})
 
-			_, err := instance.GetUpdateInstanceOpts(context.TODO())
+			_, err := instance.getUpdateOpts(context.TODO())
 
 			assert.Error(t, err)
 			assert.ErrorContains(t, err, ".contractResourceModel")
 		},
 	)
-}
-
-func Test_adaptIpToIPResource(t *testing.T) {
-	sdkIp := publiccloud.Ip{
-		Ip: "127.0.0.1",
-	}
-
-	want := ipResourceModel{
-		IP: basetypes.NewStringValue("127.0.0.1"),
-	}
-	got := adaptIpToIPResource(sdkIp)
-
-	assert.Equal(t, want, got)
-}
-
-func Test_adaptIpDetailsToIPResource(t *testing.T) {
-	sdkIpDetails := publiccloud.IpDetails{
-		Ip: "127.0.0.1",
-	}
-
-	want := ipResourceModel{
-		IP: basetypes.NewStringValue("127.0.0.1"),
-	}
-	got := adaptIpDetailsToIPResource(sdkIpDetails)
-
-	assert.Equal(t, want, got)
-}
-
-func Test_adaptIsoToISOResource(t *testing.T) {
-	iso := publiccloud.Iso{
-		Id:   "id",
-		Name: "name",
-	}
-	got := adaptIsoToISOResource(iso)
-
-	want := isoResourceModel{
-		ID:   basetypes.NewStringValue("id"),
-		Name: basetypes.NewStringValue("name"),
-	}
-
-	assert.Equal(t, want, got)
 }